Created Exporter and moved logic into
